Moved implementations under associated structs
