+++ name-core/src/structs.rs
@@ -8,8 +8,7 @@ use std::{
 use crate::{
     constants::{
         MIPS_DATA_START_ADDR, MIPS_HEAP_START_ADDR, MIPS_STACK_END_ADDR, MIPS_TEXT_START_ADDR,
-    },
-    syscalls::*,
+    }, exception::constants::EXCEPTION_BEING_HANDLED, syscalls::*
 };
 
 /// Symbol is used for assembly -> ELF and ELF -> ProgramState construction.
@@ -40,10 +39,26 @@ impl Default for Processor {
     }
 }
 
+impl Processor {
+    pub fn new(entry: u32) -> Self {
+        Processor {
+            pc: entry,
+            general_purpose_registers: [0; 32],
+        }
+    }
+}
+
 /// Coprocessor 0 is for communication with the OS. Look in name-core/exception for more.
 #[derive(Debug, Default)]
 pub struct Coprocessor0 {
     pub registers: [u32; 32],
+}
+
+// TODO: Fill any default values for cp0 fields
+impl Coprocessor0 {
+    pub fn new() -> Self {
+        Coprocessor0 { registers: [0; 32] }
+    }
 }
 
 /// Memory is a conglomerate of program text, program data, the heap, the stack, and other segments.
@@ -259,6 +274,21 @@ pub struct ProgramState {
     pub memory: Memory,
 }
 
+impl ProgramState {
+    pub fn new(cpu: Processor, memory: Memory) -> Self {
+        ProgramState {
+            should_continue_execution: true,
+            cpu: cpu,
+            cp0: Coprocessor0::new(),
+            memory: memory,
+        }
+    }
+
+    pub fn is_exception(&self) -> bool {
+        return self.cp0.get_exception_level() == EXCEPTION_BEING_HANDLED;
+    }
+}
+
 #[derive(Debug, Clone, Copy)]
 #[repr(usize)]
 pub enum Register {
@@ -317,6 +347,19 @@ pub struct LineInfo {
     pub line_number: u32,
     pub start_address: u32,
     pub end_address: u32,
+}
+
+impl LineInfo {
+    pub fn to_bytes(&self) -> Vec<u8> {
+        let mut bytes = self.content.as_bytes().to_vec();
+        bytes.push(b'\0');
+
+        bytes.extend_from_slice(&self.line_number.to_be_bytes());
+        bytes.extend_from_slice(&self.start_address.to_be_bytes());
+        bytes.extend_from_slice(&self.end_address.to_be_bytes());
+
+        bytes
+    }
 }
 
 /// Handler for outside world. Operating System interprets syscalls.
